requirements.txt
pillow
python-dotenv
nlpcloud 